Moved the list functions to the new module
